hw/gpio: Add basic Aspeed GPIO model
This models the ast2400 and ast2500 GPIO controllers.

GPIO pins are arranged in groups of 8 pins labeled A,B,..,Y,Z,AA,AB,AC.
(Note that the ast2400 controller only goes up to group AB).
A set has four groups (except set AC which only has one) and is
referred to by the groups it is composed of (eg ABCD,EFGH,...,YZAAAB).
Each set is accessed and controlled by a bank of 14 registers.

These registers operate on a per pin level where each bit in the register
corresponds to a pin, except for the command source registers. The command
source registers operate on a per group level where bits 24, 16, 8 and 0
correspond to each group in the set.

 eg. registers for set ABCD:
 |D7...D0|C7...C0|B7...B0|A7...A0| <- GPIOs
 |31...24|23...16|15....8|7.....0| <- bit position

Note that there are a couple of groups that only have 4 pins.

There are three ways that this model deviates from the behaviour of the
actual controller:
(1) There are three debounce registers which aren't modeled and so the per
set debounce setting registers don't affect anything.

(2) The only control source driving the GPIO pins in the model is the ARM
model (as there currently aren't models for the LPC or Coprocessor [TODO]).

(3) None of the registers in the model are reset tolerant. [TODO]

Signed-off-by: Rashmica Gupta <rashmica.g@gmail.com>
[clg: - changed commit log ]
Signed-off-by: Cédric Le Goater <clg@kaod.org>