aspeed: Set cbar and don't power down secondaries
I think cbar is supposed to be set to PERIPHBASE (see
get_gicd_base_address in u-boot's arch/arm/cpu/armv7/virt-v7.c).

According to that same function the GIC is GIC_DIST_OFFSET past this
address, and we know that the GIC is at 0x40461000 according to the
ast2600 device tree.

When we combine that change with turning off the 'start powered down'
option, we can bring up both CPUs with ASPEED's aspeed-dev-v5.1 Linux
branch.

Signed-off-by: Joel Stanley <joel@jms.id.au>